safari-600.7-branch/LayoutTests/fast/events/scroll-after-click-on-tab-index-expected.txt

ALERT: Scroll position 0
Click the text below and then try to scroll.
PASS successfullyParsed is true

TEST COMPLETE
Scroll position is more than 0
PASS document.body.scrollTop > 0 is true










My span 
